requirements.txt
chromadb
langchain
langchain-community
langchain-ollama
langchain-text-splitters
langchain-huggingface
sentence-transformers
python-dotenv
requests